deployment/esp32-simulator/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
gmqtt==0.7.0
    # via simulator (pyproject.toml)
python-dotenv==1.1.1
    # via simulator (pyproject.toml)
